% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_with_progress}
\alias{download_with_progress}
\title{Download a file with simple progress indicator}
\usage{
download_with_progress(
  url,
  filename,
  platform,
  community = NULL,
  expected_size
)
}
\arguments{
\item{url}{URL to download from}

\item{filename}{Destination filename}

\item{platform}{Platform name}

\item{community}{Community name (optional)}

\item{expected_size}{Expected file size in bytes}
}
\description{
Download a file with simple progress indicator
}
\keyword{internal}
