% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{FILE_URLS}
\alias{FILE_URLS}
\title{URLs and constants for MADOC dataset}
\format{
An object of class \code{list} of length 4.
}
\usage{
FILE_URLS
}
\description{
URLs and constants for MADOC dataset
}
\keyword{internal}
