% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{format_size}
\alias{format_size}
\title{Format file size in human readable format}
\usage{
format_size(size)
}
\arguments{
\item{size}{Size in bytes}
}
\value{
Character string with formatted size
}
\description{
Format file size in human readable format
}
\keyword{internal}
