% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{format_time}
\alias{format_time}
\title{Get human readable time estimate}
\usage{
format_time(bytes_per_sec, total_bytes)
}
\arguments{
\item{bytes_per_sec}{Download speed in bytes per second}

\item{total_bytes}{Total file size in bytes}
}
\value{
Character string with time estimate
}
\description{
Get human readable time estimate
}
\keyword{internal}
