% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_available_data}
\alias{list_available_data}
\title{List available platforms and communities in the MADOC dataset}
\usage{
list_available_data()
}
\value{
Prints formatted list of available data and returns invisible list of platforms and communities
}
\description{
List available platforms and communities in the MADOC dataset
}
