% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{print_progress}
\alias{print_progress}
\title{Print download progress with size and speed information}
\usage{
print_progress(
  platform,
  community = NULL,
  downloaded_bytes,
  total_bytes,
  bytes_per_sec
)
}
\arguments{
\item{platform}{Platform name}

\item{community}{Community name (optional)}

\item{downloaded_bytes}{Bytes downloaded so far}

\item{total_bytes}{Total file size}

\item{bytes_per_sec}{Download speed in bytes/sec}
}
\description{
Print download progress with size and speed information
}
\keyword{internal}
