% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{print_success}
\alias{print_success}
\title{Print success message after download}
\usage{
print_success(filepath, size)
}
\arguments{
\item{filepath}{Full path to downloaded file}

\item{size}{File size in bytes}
}
\description{
Print success message after download
}
\keyword{internal}
